requirements.txt
# requirements.txt
gradio
tensorflow-macos==2.16.1 
tensorflow-metal==1.1.0
transformers
sentencepiece